refactor: isolate Select stories
